feat(AppSlide): Add new AppSlide component for the AppSlider component
