+++ src/components/AppQRCode.vue
@@ -81,8 +81,6 @@ const onQrDataChanged = (newValue: AppQRCodeProps) => {
     throw new Error('qrContainerEl is null!');
   }
 
-  console.debug('onQrDataChanged', newValue);
-
   const qr = qrCode(newValue.typeNumber, newValue.correctionLevel);
   qr.addData(newValue.qrData);
   qr.make();
