+++ src/components/slider/AppSlide.vue
@@ -1,0 +1,9 @@
+<!--
+  # AppSlide
+-->
+
+<template>
+  <div class="flex-none w-full">
+    <slot></slot>
+  </div>
+</template>
