+++ src/components/slider/slider.interface.ts
@@ -1,0 +1,8 @@
+export interface AppSliderProps {
+  infinite: boolean;
+}
+
+export interface AppSliderSlideEventDetails {
+  slideNumber: number;
+  slideEl: HTMLElement;
+}
